requirements.txt
hexbytes>=0.1.0,<0.4.0
web3==6.15.1
eth-account==0.10.0
aiohttp==3.9.1
python-dotenv==1.0.0
numpy==1.24.3
cupy-cuda11x==12.3.0  # Optional, for GPU support
asyncio==3.4.3
eth-abi==4.2.1
websockets==12.0
